refactor: add data usecases protocol files to simplify imports
