+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:59912/","mongoDBName":"jest"}
+{"mongoUri":"mongodb://127.0.0.1:60155/","mongoDBName":"jest"}
