+++ src/data/usecases/load-account-by-token/db-load-account-by-token-protocols.ts
@@ -1,0 +1,4 @@
+export * from '@data/protocols/cryptography/decrypter'
+export * from '@data/protocols/db/account/load-account-by-token-repository copy'
+export * from '../add-account/db-add-account-protocols'
+export * from './db-load-account-by-token'
