+++ src/data/usecases/load-account-by-token/db-load-account-by-token.spec.ts
@@ -1,7 +1,5 @@
-import { Decrypter } from '@data/protocols/cryptography/decrypter'
-import { LoadAccountByTokenRepository } from '@data/protocols/db/account/load-account-by-token-repository copy'
-import { AccountModel } from '../add-account/db-add-account-protocols'
 import { DbLoadAccountByToken } from './db-load-account-by-token'
+import { AccountModel, Decrypter, LoadAccountByTokenRepository } from './db-load-account-by-token-protocols'
 
 const makeFakeAccount = (): AccountModel => ({
   id: 'valid_id',
