+++ src/data/usecases/load-account-by-token/db-load-account-by-token.ts
@@ -1,7 +1,5 @@
-import { Decrypter } from '@data/protocols/cryptography/decrypter'
-import { LoadAccountByTokenRepository } from '@data/protocols/db/account/load-account-by-token-repository copy'
 import { LoadAccountByToken } from '@domain/usecases/load-account-by-token'
-import { AccountModel } from '../add-account/db-add-account-protocols'
+import { Decrypter, LoadAccountByTokenRepository, AccountModel } from './db-load-account-by-token-protocols'
 
 export class DbLoadAccountByToken implements LoadAccountByToken {
   constructor(
