+++ src/data/usecases/load-survey-by-id/db-load-survey-by-id-protocols.ts
@@ -1,0 +1,3 @@
+export * from '@data/protocols/db/survey/db-load-survey-by-id-repository'
+export * from '@domain/models/survey'
+export * from './db-load-survey-by-id'
