+++ src/data/usecases/load-survey-by-id/db-load-survey-by-id.ts
@@ -1,6 +1,5 @@
-import { LoadSurveyByIdRepository } from '@data/protocols/db/survey/db-load-survey-by-id-repository'
-import { SurveyModel } from '@domain/models/survey'
 import { LoadSurveyById } from '@domain/usecases/load-survey-by-id'
+import { LoadSurveyByIdRepository, SurveyModel } from './db-load-survey-by-id-protocols'
 
 export class DbLoadSurveyById implements LoadSurveyById {
   constructor(private readonly loadSurveyByIdRepository: LoadSurveyByIdRepository) {}
