+++ src/data/usecases/load-surveys/db-load-surveys.ts
@@ -1,6 +1,4 @@
-import { LoadSurveysRepository } from '@data/protocols/db/survey/db-load-survey-repository'
-import { SurveyModel } from '@domain/models/survey'
-import { LoadSurveys } from '@domain/usecases/load-surveys'
+import { LoadSurveys, LoadSurveysRepository, SurveyModel } from './db-load-surveys-protocols'
 
 export class DbLoadSurveys implements LoadSurveys {
   constructor(private readonly loadSurveysRepository: LoadSurveysRepository) {}
